LayoutTests/fast/dom/HTMLSelectElement/selected-index-preserved-when-option-text-changes-expected.txt


Fix for bug 70547-- changing the text of an option in a select element with no selected option (-1) should preserve that option

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Running tests on menu list
Setting selected index to -1
Changed text attribute of first item
PASS mySelect.selectedIndex is -1
Changed value attribute of first item
PASS mySelect.selectedIndex is -1
Changed label attribute of first item
PASS mySelect.selectedIndex is -1
Setting selected index to 1
Changed text attribute of first item
PASS mySelect.selectedIndex is 1
Changed value attribute of first item
PASS mySelect.selectedIndex is 1
Changed label attribute of first item
PASS mySelect.selectedIndex is 1
Running tests on list box
Setting selected index to -1
Changed text attribute of first item
PASS mySelect.selectedIndex is -1
Changed value attribute of first item
PASS mySelect.selectedIndex is -1
Changed label attribute of first item
PASS mySelect.selectedIndex is -1
Setting selected index to 1
Changed text attribute of first item
PASS mySelect.selectedIndex is 1
Changed value attribute of first item
PASS mySelect.selectedIndex is 1
Changed label attribute of first item
PASS mySelect.selectedIndex is 1
PASS successfullyParsed is true

TEST COMPLETE
